Log information about deleted artifacts, strict JobID filter
Check that job_status is both defined and non-empty to avoid deleting containers that partially match a jobID in the name

TestBos2Sno: sno

+++ utils/cleanup-scripts/clean-webartifact.sh
@@ -19,6 +19,8 @@
 # The Remore CI is used to confirm the Job status
 # Params: The path to a valid remote CI
 
+exec >> "$HOME"/containers-cleanup.log 2>&1
+
 function help(){
     echo "$(basename "$0") <remote_ci_path>"
     echo "Absolute path to remote CI is missing"
@@ -30,6 +32,9 @@ if [[ $# != 1 ]]; then
     exit 1
 fi
 
+echo "================================================================================"
+date
+
 # Source the remote CI file
 remote_ci="${1}"
 if [[ ! -f $remote_ci ]]; then
@@ -39,11 +44,11 @@ fi
 source "${remote_ci}"
 
 ###
-# Cleaning web artifacts
+# Cleaning web-artifacts containers
 ###
 
-# Get the container names
-containers=$(podman ps -a --sort created --filter 'name=\w{8}(-\w{4}){3}-\w{8}' --format "{{.Names}}");
+# Get the web-container names
+containers=$(podman ps -a --sort created --filter 'name=^\w{8}(-\w{4}){3}-\w{12}$' --format "{{.Names}}");
 
 # Loop over the containers and check their job status
 while IFS=, read -r name
@@ -51,10 +56,11 @@ do
   # Check if the job is in a failure state
   fail_states=( failure error killed )
   job_status=$(dcictl --format json job-show "${name}" | jq -er .job.status)
-  if [[ " ${fail_states[*]} " =~ ${job_status} ]]; then
+  if [[ -n "$job_status" && " ${fail_states[*]} " =~ ${job_status} ]]; then
     port=$(podman inspect "${name}" | jq -r '.[].NetworkSettings.Ports."8080/tcp"[0].HostPort')
     sudo firewall-cmd --remove-port="${port}"/tcp
     podman rm -f "${name}"
+    echo "Removed container ${name}"
   fi
 done <<< "${containers}"
 
@@ -82,5 +88,6 @@ do
   job_status=$(dcictl --format json job-show "${job_id}" | jq -er .job.status)
   if [[ " ${fail_states[*]} " =~ ${job_status} ]]; then
     podman rmi -f "${name}"
+    echo "Removed image ${name}"
   fi
 done < /tmp/clean-images
